[BE] refactor: 레디스 맵 버퍼 클래스 분리  #763
