+++ backend/app/src/main/java/com/onair/hearit/app/playinghistory/infrastructure/buffer/PlayingHistoryBuffer.java
@@ -7,4 +7,6 @@ public interface PlayingHistoryBuffer {
     void add(PlayingHistory playingHistory, long clientEventTime);
 
     void flush();
+
+    int size();
 }
